fixed a "bug" with interpolation of input.
If you omit the value at time point 1 and/or the end of the simulation for 'I' and 'noise', than input was not defined for the start and/or end. Therefore it was set to NaN, which resulted in breaking of the simulation.

Now, it assumes undefined input to be equal to 0.